Initial commit for Julia-JLLUtils
Current utils:

`nanmean` - computes mean along array dimension or whole array ignoring
NaNs.

`move_dim` - rearrange an array so that a particular dimension is first,
or in a specific position.

Exceptions: set up a framework for simple Python-like exceptions that
accept a single argument, the message to print. For now, exceptions
created as subtypes of the `ExceptionWithMsg` abstract type will have
to remember to add a field for their msg. Such custom exceptions can
be created with the `@msgexc` macro.